Create Melisa.md
# Steganography Img
To create an encrypted image, you must first upload 2 images of a large enough size from internet.

Then download OpenStego software, extract the zip and install Java.

Open OpenStego and put the secret in "message file" and the image that hides it in "cover file". And choose the value in "Output Stego File" to indicate the name and destination of the final result. Finally click on "hide data".

To decrypt it, go to "extract data" on the left.

Then put the result in the "image" folder of the group's branch, and make a "markdown" with a link to the image.

Click on "commit" and then update the content in github by clicking on "fetch origin" for adding the work.

Finally, make a markdown in Atom, in "pages", create a "new file" and name it with ".md" at the end.

![To Decrypt](../image/images2.bmp)